feat : 상품-태그 연동 기능 개발#1

+++ src/main/java/com/nhnacademy/booklay/server/entity/ProductTag.java
@@ -31,10 +31,10 @@ public class ProductTag {
     public static class Pk implements Serializable {
 
         @Column(name = "product_no", nullable = false)
-        private Integer productId;
+        private Long productId;
 
         @Column(name = "tag_no", nullable = false)
-        private Integer tagId;
+        private Long tagId;
 
     }
